requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
numpy
beautifulsoup4
requests
python-multipart
torch
pandas